Minor changes Pagelist
